move all issue tests to regression_test.go
